app/requirements.txt
Django>=4.2.0
# groq>=0.4.0
requests>=2.31.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0
whitenoise>=6.6.0
celery[redis]>=5.3.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0
redis>=5.0.0
PyJWT>=2.8.0
youtube-search
